Fix UnicodeDecodeError raise on av.open()
This happens because the encoding of the metadata by default
is assumed to be UTF-8, but some of the files from the FFmpeg
test suite actually use latin1.

As metadata is not heavily used in the project, still default
to UTF-8 (which is quite sensible), but simply ignore
non-conformant characters.